move new strategy change into the right firle
